App Structure
build: build app structure with Viber
build network manager with moya
build main screen design with custom fonts and assets
build launch screen

+++ WeatherIOSApp/Entity/WeatherForecastResponse/Hour.swift
@@ -67,7 +67,7 @@ struct Hour: Codable {
     var precipMm: Float?
     var windDir: String?
     var heatindexF: Float?
-    var windKph: Int?
+    var windKph: Float?
     var isDay: Int?
     var tempC: Float?
     var cloud: Int?
